add zod.d.ts
